Support this project
Android studio project to create an app that connects to a remote DB. The project makes use of the Volley library for HTTP connection, so the permissions must be configured in the manifest.xml and import the library in the project dependencies, to later import it as a Class in the .java files. Webservices in php are used to process the SQL query to a MySQL database and the result is retrieved through a JSON array, which is processed in the app through objects of the User(POJO) class. All the project files have been verified and the project is fully functional as of September 2022. However, I recommend the following video from my channel, in which I explain the process step by step: https://youtu.be/EzRZVsw70Rk

Although nowadays there are many frameworks that do all this work; it is good to understand how to do it manually. I hope you find this project useful.

Best regards,